requirements.txt
Flask
Pillow
requests
gunicorn # Servidor web para producción